requirements.txt
asgiref==3.4.1
Django==3.2.9
django-debug-toolbar==3.2.2
django-seed==0.3.1
djangorestframework==3.12.4
Faker==9.8.2
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
